Correct the type on sessionIdentifierMetadata.
This has never been limited to `NSString` values, and if it was called from
Swift it likely would cause problems. Luckily, it is really meant to be an
internal api so clients shouldn't be calling it. But since it is public, it is a
breaking change to Swift.

Part of #398